🔨 Migrate to provide/inject rather than prop piping
