+++ src/sections/FoldersSection.vue
@@ -16,25 +16,19 @@
 </template>
 
 <script setup>
-import { ref, toRef } from 'vue';
+import { ref, inject } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { tabsFolders } from '@/definitions.js';
 import { useFoldersChartData } from '@/composables/useFoldersChartData.js';
 import DoughnutChart from '@/charts/DoughnutChart.vue';
 import SectionTabHeader from '@/parts/SectionTabHeader.vue';
 
-const props = defineProps({
-	display: { type: Object, required: true },
-});
+const { display } = inject('engine');
 
 const { t, locale } = useI18n();
 
 // tab navigation local to this section (single-entry tab enum, kept for template consistency)
 const tabFolders = ref(tabsFolders.foldersDistribution);
 
-const { foldersChartData } = useFoldersChartData({
-	display: toRef(props, 'display'),
-	locale,
-	t,
-});
+const { foldersChartData } = useFoldersChartData({ display, locale, t });
 </script>
