+++ src/sections/TotalSection.vue
@@ -86,33 +86,36 @@
 </template>
 
 <script setup>
-import { ref, toRef } from 'vue';
+import { ref, computed, inject } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { tabsTotal } from '@/definitions.js';
 import { useTotalChartData } from '@/composables/useTotalChartData.js';
 import LineChart from '@/charts/LineChart.vue';
 import SectionTabHeader from '@/parts/SectionTabHeader.vue';
 
-const props = defineProps({
-	display: { type: Object, required: true },
-	comparison: { type: Object, required: true },
-	accounts: { type: Array, required: true },
-	options: { type: Object, required: true },
-	minYear: { type: Number, required: true },
-	maxYear: { type: Number, required: true },
-	minDate: { type: Date, required: true },
-	maxDate: { type: Date, required: true },
-	preferences: { type: Object, required: true },
-	singleAccount: { type: Boolean, required: true },
-	accountsColorGradient: { type: String, required: true },
-	tooltipAccountComparison: { type: Function, required: true },
-	unfinished: { type: Boolean, required: true },
-});
+const {
+	display,
+	comparison,
+	accounts,
+	options,
+	minYear,
+	maxYear,
+	minDate,
+	maxDate,
+	preferences,
+	singleAccount,
+	accountsColorGradient,
+	tooltipAccountComparison,
+	active,
+} = inject('engine');
 
 const { t, locale } = useI18n();
 
 // tab navigation local to this section
 const tabTotal = ref(tabsTotal.years);
+
+// true while the current period has no end date set (chart's last segment is still "in progress")
+const unfinished = computed(() => active.period.end == null);
 
 const {
 	yearsChartData,
@@ -123,16 +126,5 @@ const {
 	monthsComparedChartData,
 	weeksChartData,
 	weeksComparedChartData,
-} = useTotalChartData({
-	display: toRef(props, 'display'),
-	comparison: toRef(props, 'comparison'),
-	accounts: toRef(props, 'accounts'),
-	options: props.options,
-	minYear: toRef(props, 'minYear'),
-	maxYear: toRef(props, 'maxYear'),
-	minDate: toRef(props, 'minDate'),
-	maxDate: toRef(props, 'maxDate'),
-	t,
-	locale,
-});
+} = useTotalChartData({ display, comparison, accounts, options, minYear, maxYear, minDate, maxDate, t, locale });
 </script>
