+++ src/sections/TwodimSection.vue
@@ -27,24 +27,19 @@
 </template>
 
 <script setup>
-import { ref, toRef } from 'vue';
+import { ref, inject } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { tabsTwodim } from '@/definitions.js';
 import { useTwodimChartData } from '@/composables/useTwodimChartData.js';
 import MatrixChart from '@/charts/MatrixChart.vue';
 import SectionTabHeader from '@/parts/SectionTabHeader.vue';
 
-const props = defineProps({
-	display: { type: Object, required: true },
-});
+const { display } = inject('engine');
 
 const { t } = useI18n();
 
 // tab navigation local to this section (single-entry tab enum, kept for template consistency)
 const tabTwodim = ref(tabsTwodim.temporalDistribution);
 
-const { weekdayPerHourChartData } = useTwodimChartData({
-	display: toRef(props, 'display'),
-	t,
-});
+const { weekdayPerHourChartData } = useTwodimChartData({ display, t });
 </script>
